[Minor-Bug] Tokens for deleted users were not cleaned up (#669)
When a user was deleted, even if they couldn't log in, token remained
valid even though they had no permission to do anything (all their
groups were removed).

This change ensures they they will get bounced at the first call,
their existing auth tokens not even being valid any further.

No security risk associated with this change, it is simply a cleanup.

+++ backend/mr/src/main/java/io/featurehub/mr/resources/PersonResource.java
@@ -126,7 +126,9 @@ public class PersonResource implements PersonServiceDelegate {
     if (authManager.isOrgAdmin(authManager.from(securityContext))) {
       Person p = getPerson(id, false, false, securityContext);
 
-      return personApi.delete(p.getEmail());
+      if (p.getEmail() != null) {
+        return personApi.delete(p.getEmail());
+      }
     }
 
     throw new ForbiddenException("No permission");
